requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy>=1.21.0
torch>=2.0.1
transformers>=4.33.0
azureml-core>=1.53.0
azure-ai-ml==1.10.0
python-dotenv>=1.0.0
